requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jieba==0.42.1
jiter==0.8.2
lxml==5.3.0
openai==1.61.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
python-http-client==3.3.7
python-pptx==1.0.2
requests==2.32.3
sendgrid==6.11.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
XlsxWriter==3.2.2
